+++ include/world_builder/features/continental_plate_models/composition/uniform.h
@@ -23,12 +23,14 @@
 
 #include "world_builder/features/continental_plate_models/composition/interface.h"
 #include "world_builder/objects/surface.h"
+#include "world_builder/features/feature_utilities.h"
 
 
 namespace WorldBuilder
 {
   namespace Features
   {
+    using namespace FeatureUtilities;
     namespace ContinentalPlateModels
     {
       namespace Composition
@@ -86,7 +88,7 @@ namespace WorldBuilder
             Objects::Surface max_depth_surface;
             std::vector<unsigned int> compositions;
             std::vector<double> fractions;
-            std::string operation;
+            Operations operation;
 
         };
       } // namespace Composition