+++ include/world_builder/features/fault_models/temperature/interface.h
@@ -23,7 +23,7 @@
 #include "world_builder/grains.h"
 #include "world_builder/parameters.h"
 #include "world_builder/utilities.h"
-#include "world_builder/features/utilities.h"
+#include "world_builder/features/feature_utilities.h"
 
 #include <map>
 #include <vector>
@@ -40,7 +40,7 @@ namespace WorldBuilder
    */
   namespace Features
   {
-
+    using namespace FeatureUtilities;
     namespace FaultModels
     {
       namespace Temperature
@@ -87,7 +87,7 @@ namespace WorldBuilder
                                    const double feature_min_depth,
                                    const double feature_max_depth,
                                    const WorldBuilder::Utilities::PointDistanceFromCurvedPlanes &distance_from_planes,
-                                   const Utilities::AdditionalParameters &additional_paramters) const = 0;
+                                   const AdditionalParameters &additional_paramters) const = 0;
             /**
              * A function to register a new type. This is part of the automatic
              * registration of the object factory.