+++ include/world_builder/features/oceanic_plate_models/temperature/adiabatic.h
@@ -22,15 +22,15 @@
 
 
 #include "world_builder/features/oceanic_plate_models/temperature/interface.h"
-#include "world_builder/features/utilities.h"
+#include "world_builder/features/feature_utilities.h"
 #include "world_builder/objects/surface.h"
 
 
 namespace WorldBuilder
 {
-
   namespace Features
   {
+    using namespace FeatureUtilities;
     namespace OceanicPlateModels
     {
       namespace Temperature
@@ -101,7 +101,7 @@ namespace WorldBuilder
              */
             double specific_heat;
 
-            Utilities::Operations operation;
+            Operations operation;
 
         };
       } // namespace Temperature