+++ include/world_builder/features/oceanic_plate_models/temperature/half_space_model.h
@@ -22,7 +22,7 @@
 
 
 #include "world_builder/features/oceanic_plate_models/temperature/interface.h"
-#include "world_builder/features/utilities.h"
+#include "world_builder/features/feature_utilities.h"
 #include "world_builder/objects/surface.h"
 
 
@@ -33,6 +33,7 @@ namespace WorldBuilder
 
   namespace Features
   {
+    using namespace FeatureUtilities;
     namespace OceanicPlateModels
     {
       namespace Temperature
@@ -91,7 +92,7 @@ namespace WorldBuilder
             double bottom_temperature;
             double spreading_velocity;
             std::vector<std::vector<Point<2> > > mid_oceanic_ridges;
-            Utilities::Operations operation;
+            Operations operation;
 
         };
       } // namespace Temperature