+++ include/world_builder/features/subducting_plate_models/temperature/mass_conserving.h
@@ -22,7 +22,7 @@
 
 
 #include "world_builder/features/subducting_plate_models/temperature/interface.h"
-#include "world_builder/features/utilities.h"
+#include "world_builder/features/feature_utilities.h"
 
 
 namespace WorldBuilder
@@ -89,7 +89,7 @@ namespace WorldBuilder
                                    const double feature_min_depth,
                                    const double feature_max_depth,
                                    const WorldBuilder::Utilities::PointDistanceFromCurvedPlanes &distance_from_planes,
-                                   const Utilities::AdditionalParameters &additional_paramters) const override final;
+                                   const AdditionalParameters &additional_paramters) const override final;
 
 
           private:
@@ -109,7 +109,7 @@ namespace WorldBuilder
             double taper_distance;
             bool adiabatic_heating;
             std::vector<std::vector<Point<2>>> mid_oceanic_ridges;
-            Utilities::Operations operation;
+            Operations operation;
 
         };
       } // namespace Temperature