+++ source/world_builder/features/oceanic_plate_models/temperature/uniform.cc
@@ -46,7 +46,7 @@ namespace WorldBuilder
           min_depth(NaN::DSNAN),
           max_depth(NaN::DSNAN),
           temperature(NaN::DSNAN),
-          operation(Utilities::Operations::REPLACE)
+          operation(Operations::REPLACE)
         {
           this->world = world_;
           this->name = "uniform";
@@ -83,7 +83,7 @@ namespace WorldBuilder
           min_depth = min_depth_surface.minimum;
           max_depth_surface = Objects::Surface(prm.get("max depth",coordinates));
           max_depth = max_depth_surface.maximum;
-          operation = Utilities::string_operations_to_enum(prm.get<std::string>("operation"));
+          operation = string_operations_to_enum(prm.get<std::string>("operation"));
           temperature = prm.get<double>("temperature");
         }
 
@@ -104,7 +104,7 @@ namespace WorldBuilder
               const double max_depth_local = max_depth_surface.constant_value ? max_depth : max_depth_surface.local_value(position_in_natural_coordinates.get_surface_point()).interpolated_value;
               if (depth <= max_depth_local &&  depth >= min_depth_local)
                 {
-                  return Utilities::apply_operation(operation,temperature_,temperature);
+                  return apply_operation(operation,temperature_,temperature);
                 }
             }