+++ source/world_builder/features/subducting_plate_models/temperature/plate_model.cc
@@ -48,7 +48,7 @@ namespace WorldBuilder
           potential_mantle_temperature(NaN::DSNAN),
           surface_temperature(NaN::DSNAN),
           adiabatic_heating(true),
-          operation(Utilities::Operations::REPLACE)
+          operation(Operations::REPLACE)
         {
           this->world = world_;
           this->name = "plate model";
@@ -101,7 +101,7 @@ namespace WorldBuilder
 
           min_depth = prm.get<double>("min distance slab top");
           max_depth = prm.get<double>("max distance slab top");
-          operation = Utilities::string_operations_to_enum(prm.get<std::string>("operation"));
+          operation = string_operations_to_enum(prm.get<std::string>("operation"));
 
           density = prm.get<double>("density");
           plate_velocity = prm.get<double>("plate velocity");
@@ -135,7 +135,7 @@ namespace WorldBuilder
                                     const double /*feature_min_depth*/,
                                     const double /*feature_max_depth*/,
                                     const WorldBuilder::Utilities::PointDistanceFromCurvedPlanes &distance_from_planes,
-                                    const Utilities::AdditionalParameters &additional_paramters) const
+                                    const AdditionalParameters &additional_paramters) const
         {
           const double thickness_local = std::min(additional_paramters.local_thickness, max_depth);
           const double distance_from_plane = distance_from_planes.distance_from_plane;
@@ -200,7 +200,7 @@ namespace WorldBuilder
               WBAssert(std::isfinite(temperature), "Internal error: temperature is not finite: " << temperature << '.');
 
 
-              return Utilities::apply_operation(operation,temperature_,temperature);
+              return apply_operation(operation,temperature_,temperature);
             }
 
           return temperature_;